@@ -2,11 +2,10 @@
 // Use of this source code is governed by a the MIT license that can be
 // found in the LICENSE file.
 
+import 'dart:math' show min, max;
+
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
-import 'package:meta/meta.dart';
-
-import 'dart:math' show min, max;
 
 /// Called every layout to provide the amount of stickyness a header is in.
 /// This lets the widgets animate their content and provide feedback.
@@ -24,40 +23,37 @@ class RenderStickyHeader extends RenderBox
     with
         ContainerRenderObjectMixin<RenderBox, MultiChildLayoutParentData>,
         RenderBoxContainerDefaultsMixin<RenderBox, MultiChildLayoutParentData> {
-
-  RenderStickyHeaderCallback _callback;
-  ScrollableState _scrollable;
+  RenderStickyHeaderCallback? _callback;
+  ScrollPosition _scrollPosition;
   bool _overlapHeaders;
 
   RenderStickyHeader({
-    @required ScrollableState scrollable,
-    RenderStickyHeaderCallback callback,
+    required ScrollPosition scrollPosition,
+    RenderStickyHeaderCallback? callback,
     bool overlapHeaders: false,
-    RenderBox header,
-    RenderBox content,
-  })  : assert(scrollable != null),
-        _scrollable = scrollable,
+    RenderBox? header,
+    RenderBox? content,
+  })  : _scrollPosition = scrollPosition,
         _callback = callback,
         _overlapHeaders = overlapHeaders {
     if (content != null) add(content);
     if (header != null) add(header);
   }
 
-  set scrollable(ScrollableState newValue) {
-    assert(newValue != null);
-    if (_scrollable == newValue) {
+  set scrollPosition(ScrollPosition newValue) {
+    if (_scrollPosition == newValue) {
       return;
     }
-    final ScrollableState oldValue = _scrollable;
-    _scrollable = newValue;
+    final ScrollPosition oldValue = _scrollPosition;
+    _scrollPosition = newValue;
     markNeedsLayout();
     if (attached) {
-      oldValue.position?.removeListener(markNeedsLayout);
-      newValue.position?.addListener(markNeedsLayout);
+      oldValue.removeListener(markNeedsLayout);
+      newValue.addListener(markNeedsLayout);
     }
   }
 
-  set callback(RenderStickyHeaderCallback newValue) {
+  set callback(RenderStickyHeaderCallback? newValue) {
     if (_callback == newValue) {
       return;
     }
@@ -76,19 +72,19 @@ class RenderStickyHeader extends RenderBox
   @override
   void attach(PipelineOwner owner) {
     super.attach(owner);
-    _scrollable.position?.addListener(markNeedsLayout);
+    _scrollPosition.addListener(markNeedsLayout);
   }
 
   @override
   void detach() {
-    _scrollable.position?.removeListener(markNeedsLayout);
+    _scrollPosition.removeListener(markNeedsLayout);
     super.detach();
   }
 
   // short-hand to access the child RenderObjects
-  RenderBox get _headerBox => lastChild;
+  RenderBox get _headerBox => lastChild!;
 
-  RenderBox get _contentBox => firstChild;
+  RenderBox get _contentBox => firstChild!;
 
   @override
   void performLayout() {
@@ -105,16 +101,16 @@ class RenderStickyHeader extends RenderBox
 
     // determine size of ourselves based on content widget
     final width = max(constraints.minWidth, _contentBox.size.width);
-    final height = max(constraints.minHeight,
-        _overlapHeaders ? contentHeight : headerHeight + contentHeight);
-    size = new Size(width, height);
+    final height =
+        max(constraints.minHeight, _overlapHeaders ? contentHeight : headerHeight + contentHeight);
+    size = Size(width, height);
     assert(size.width == constraints.constrainWidth(width));
     assert(size.height == constraints.constrainHeight(height));
     assert(size.isFinite);
 
     // place content underneath header
     final contentParentData = _contentBox.parentData as MultiChildLayoutParentData;
-    contentParentData.offset = new Offset(0.0, _overlapHeaders ? 0.0 : headerHeight);
+    contentParentData.offset = Offset(0.0, _overlapHeaders ? 0.0 : headerHeight);
 
     // determine by how much the header should be stuck to the top
     final double stuckOffset = determineStuckOffset();
@@ -122,17 +118,17 @@ class RenderStickyHeader extends RenderBox
     // place header over content relative to scroll offset
     final double maxOffset = height - headerHeight;
     final headerParentData = _headerBox.parentData as MultiChildLayoutParentData;
-    headerParentData.offset = new Offset(0.0, max(0.0, min(-stuckOffset, maxOffset)));
+    headerParentData.offset = Offset(0.0, max(0.0, min(-stuckOffset, maxOffset)));
 
     // report to widget how much the header is stuck.
     if (_callback != null) {
       final stuckAmount = max(min(headerHeight, stuckOffset), -headerHeight) / headerHeight;
-      _callback(stuckAmount);
+      _callback!(stuckAmount);
     }
   }
 
   double determineStuckOffset() {
-    final scrollBox = _scrollable.context.findRenderObject();
+    final scrollBox = _scrollPosition.context.notificationContext!.findRenderObject();
     if (scrollBox?.attached ?? false) {
       try {
         return localToGlobal(Offset.zero, ancestor: scrollBox).dy;
@@ -147,7 +143,7 @@ class RenderStickyHeader extends RenderBox
   void setupParentData(RenderObject child) {
     super.setupParentData(child);
     if (child.parentData is! MultiChildLayoutParentData) {
-      child.parentData = new MultiChildLayoutParentData();
+      child.parentData = MultiChildLayoutParentData();
     }
   }
 
@@ -163,26 +159,26 @@ class RenderStickyHeader extends RenderBox
 
   @override
   double computeMinIntrinsicHeight(double width) {
-    return _overlapHeaders ? _contentBox.getMinIntrinsicHeight(width)
-        : (_headerBox.getMinIntrinsicHeight(width) +
-            _contentBox.getMinIntrinsicHeight(width));
+    return _overlapHeaders
+        ? _contentBox.getMinIntrinsicHeight(width)
+        : (_headerBox.getMinIntrinsicHeight(width) + _contentBox.getMinIntrinsicHeight(width));
   }
 
   @override
   double computeMaxIntrinsicHeight(double width) {
-    return _overlapHeaders ? _contentBox.getMaxIntrinsicHeight(width)
-        : (_headerBox.getMaxIntrinsicHeight(width) +
-            _contentBox.getMaxIntrinsicHeight(width));
+    return _overlapHeaders
+        ? _contentBox.getMaxIntrinsicHeight(width)
+        : (_headerBox.getMaxIntrinsicHeight(width) + _contentBox.getMaxIntrinsicHeight(width));
   }
 
   @override
-  double computeDistanceToActualBaseline(TextBaseline baseline) {
+  double? computeDistanceToActualBaseline(TextBaseline baseline) {
     return defaultComputeDistanceToHighestActualBaseline(baseline);
   }
 
   @override
-  bool hitTestChildren(HitTestResult result, {Offset position}) {
-    return defaultHitTestChildren(result, position: position);
+  bool hitTestChildren(HitTestResult result, {required Offset position}) {
+    return defaultHitTestChildren(result as BoxHitTestResult, position: position);
   }
 
   @override
