@@ -3,11 +3,11 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:meta/meta.dart';
+
 import './render.dart';
 
 /// Builder called during layout to allow the header's content to be animated or styled based
-/// on the amount of stickyness the header has.
+/// on the amount of stickiness the header has.
 ///
 /// [context] for your build operation.
 ///
@@ -30,10 +30,11 @@ typedef Widget StickyHeaderWidgetBuilder(BuildContext context, double stuckAmoun
 class StickyHeader extends MultiChildRenderObjectWidget {
   /// Constructs a new [StickyHeader] widget.
   StickyHeader({
-    Key key,
-    @required this.header,
-    @required this.content,
+    Key? key,
+    required this.header,
+    required this.content,
     this.overlapHeaders: false,
+    this.controller,
     this.callback,
   }) : super(
           key: key,
@@ -50,16 +51,18 @@ class StickyHeader extends MultiChildRenderObjectWidget {
   /// If true, the header will overlap the Content.
   final bool overlapHeaders;
 
-  /// Optional callback with stickyness value. If you think you need this, then you might want to
+  /// Optional [ScrollController] that will be used by the widget instead of the default inherited one.
+  final ScrollController? controller;
+
+  /// Optional callback with stickiness value. If you think you need this, then you might want to
   /// consider using [StickyHeaderBuilder] instead.
-  final RenderStickyHeaderCallback callback;
+  final RenderStickyHeaderCallback? callback;
 
   @override
   RenderStickyHeader createRenderObject(BuildContext context) {
-    var scrollable = Scrollable.of(context);
-    assert(scrollable != null);
-    return new RenderStickyHeader(
-      scrollable: scrollable,
+    final scrollPosition = this.controller?.position ?? Scrollable.of(context)!.position;
+    return RenderStickyHeader(
+      scrollPosition: scrollPosition,
       callback: this.callback,
       overlapHeaders: this.overlapHeaders,
     );
@@ -67,8 +70,9 @@ class StickyHeader extends MultiChildRenderObjectWidget {
 
   @override
   void updateRenderObject(BuildContext context, RenderStickyHeader renderObject) {
+    final scrollPosition = this.controller?.position ?? Scrollable.of(context)!.position;
     renderObject
-      ..scrollable = Scrollable.of(context)
+      ..scrollPosition = scrollPosition
       ..callback = this.callback
       ..overlapHeaders = this.overlapHeaders;
   }
@@ -84,10 +88,11 @@ class StickyHeader extends MultiChildRenderObjectWidget {
 class StickyHeaderBuilder extends StatefulWidget {
   /// Constructs a new [StickyHeaderBuilder] widget.
   const StickyHeaderBuilder({
-    Key key,
-    @required this.builder,
-    this.content,
+    Key? key,
+    required this.builder,
+    required this.content,
     this.overlapHeaders: false,
+    this.controller,
   }) : super(key: key);
 
   /// Called when the sticky amount changes for the header.
@@ -100,26 +105,30 @@ class StickyHeaderBuilder extends StatefulWidget {
   /// If true, the header will overlap the Content.
   final bool overlapHeaders;
 
+  /// Optional [ScrollController] that will be used by the widget instead of the default inherited one.
+  final ScrollController? controller;
+
   @override
-  _StickyHeaderBuilderState createState() => new _StickyHeaderBuilderState();
+  _StickyHeaderBuilderState createState() => _StickyHeaderBuilderState();
 }
 
 class _StickyHeaderBuilderState extends State<StickyHeaderBuilder> {
-  double _stuckAmount;
+  double? _stuckAmount;
 
   @override
   Widget build(BuildContext context) {
-    return new StickyHeader(
+    return StickyHeader(
       overlapHeaders: widget.overlapHeaders,
-      header: new LayoutBuilder(
+      header: LayoutBuilder(
         builder: (context, _) => widget.builder(context, _stuckAmount ?? 0.0),
       ),
       content: widget.content,
+      controller: widget.controller,
       callback: (double stuckAmount) {
         if (_stuckAmount != stuckAmount) {
           _stuckAmount = stuckAmount;
-          WidgetsBinding.instance.endOfFrame.then((_) {
-            if(mounted){
+          WidgetsBinding.instance!.endOfFrame.then((_) {
+            if (mounted) {
               setState(() {});
             }
           });
