@@ -1,205 +1,308 @@
 import 'package:flutter/material.dart';
-import 'package:meta/meta.dart';
 import 'package:sticky_headers/sticky_headers.dart';
 
 import './images.dart';
 
-void main() => runApp(new ExampleApp());
+void main() => runApp(ExampleApp());
 
 class ExampleApp extends StatelessWidget {
-	@override
-	Widget build(BuildContext context) {
-		return new MaterialApp(
-			title: 'Sticky Headers Example',
-			theme: new ThemeData(
-				primarySwatch: Colors.blueGrey,
-			),
-			home: new MainScreen(),
-		);
-	}
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Sticky Headers Example',
+      theme: ThemeData(
+        primarySwatch: Colors.blueGrey,
+      ),
+      home: MainScreen(),
+    );
+  }
 }
 
 class MainScreen extends StatelessWidget {
-	@override
-	Widget build(BuildContext context) {
-		return new ScaffoldWrapper(
-			title: 'Sticky Headers Example',
-			child: new ListView(
-				children: ListTile.divideTiles(
-					context: context,
-					tiles: <Widget>[
-						new ListTile(
-							title: const Text('Example 1 - Headers and Content'),
-							onTap: () => navigateTo(context, (context) => new Example1()),
-						),
-						new ListTile(
-							title: const Text('Example 2 - Animated Headers with Content'),
-							onTap: () => navigateTo(context, (context) => new Example2()),
-						),
-						new ListTile(
-							title: const Text('Example 3 - Headers overlapping the Content'),
-							onTap: () => navigateTo(context, (context) => new Example3()),
-						),
-					],
-				).toList(growable: false),
-			),
-		);
-	}
+  @override
+  Widget build(BuildContext context) {
+    return ScaffoldWrapper(
+      title: 'Sticky Headers Example',
+      child: ListView(
+        children: ListTile.divideTiles(
+          context: context,
+          tiles: <Widget>[
+            ListTile(
+              title: const Text('Example 1 - Headers and Content'),
+              onTap: () => navigateTo(context, (context) => Example1()),
+            ),
+            ListTile(
+              title: const Text('Example 2 - Animated Headers with Content'),
+              onTap: () => navigateTo(context, (context) => Example2()),
+            ),
+            ListTile(
+              title: const Text('Example 3 - Headers overlapping the Content'),
+              onTap: () => navigateTo(context, (context) => Example3()),
+            ),
+            ListTile(
+              title: const Text('Example 4 - Example using scroll controller'),
+              onTap: () => navigateTo(context, (context) => Example4()),
+            ),
+          ],
+        ).toList(growable: false),
+      ),
+    );
+  }
 
-	navigateTo(BuildContext context, builder(BuildContext context)) {
-		Navigator.of(context).push(new MaterialPageRoute(builder: builder));
-	}
+  void navigateTo(BuildContext context, WidgetBuilder builder) {
+    Navigator.of(context).push(MaterialPageRoute(builder: builder));
+  }
 }
 
 class Example1 extends StatelessWidget {
-	@override
-	Widget build(BuildContext context) {
-		return new ScaffoldWrapper(
-			title: 'Example 1',
-			child: new ListView.builder(itemBuilder: (context, index) {
-				return new Material(
-					color: Colors.grey[300],
-					child: new StickyHeader(
-						header: new Container(
-							height: 50.0,
-							color: Colors.blueGrey[700],
-							padding: new EdgeInsets.symmetric(horizontal: 16.0),
-							alignment: Alignment.centerLeft,
-							child: new Text('Header #$index',
-								style: const TextStyle(color: Colors.white),
-							),
-						),
-						content: new Container(
-							child: new Image.network(imageForIndex(index), fit: BoxFit.cover,
-								width: double.infinity, height: 200.0),
-						),
-					),
-				);
-			}),
-		);
-	}
+  const Example1({
+    Key? key,
+    this.controller,
+  }) : super(key: key);
 
-	String imageForIndex(int index) {
-		return Images.imageThumbUrls[index % Images.imageThumbUrls.length];
-	}
+  final ScrollController? controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return ScaffoldWrapper(
+      wrap: controller == null,
+      title: 'Example 1',
+      child: ListView.builder(
+        primary: controller == null,
+        controller: controller,
+        itemBuilder: (context, index) {
+          return Material(
+            color: Colors.grey[300],
+            child: StickyHeader(
+              controller: controller, // Optional
+              header: Container(
+                height: 50.0,
+                color: Colors.blueGrey[700],
+                padding: EdgeInsets.symmetric(horizontal: 16.0),
+                alignment: Alignment.centerLeft,
+                child: Text(
+                  'Header #$index',
+                  style: const TextStyle(color: Colors.white),
+                ),
+              ),
+              content: Container(
+                child: Image.network(imageForIndex(index),
+                    fit: BoxFit.cover, width: double.infinity, height: 200.0),
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+
+  String imageForIndex(int index) {
+    return Images.imageThumbUrls[index % Images.imageThumbUrls.length];
+  }
 }
 
 class Example2 extends StatelessWidget {
-	@override
-	Widget build(BuildContext context) {
-		return new ScaffoldWrapper(
-			title: 'Example 2',
-			child: new ListView.builder(itemBuilder: (context, index) {
-				return new Material(
-					color: Colors.grey[300],
-					child: new StickyHeaderBuilder(
-						builder: (BuildContext context, double stuckAmount) {
-							stuckAmount = 1.0 - stuckAmount.clamp(0.0, 1.0);
-							return new Container(
-								height: 50.0,
-								color: Color.lerp(Colors.blue[700], Colors.red[700], stuckAmount),
-								padding: new EdgeInsets.symmetric(horizontal: 16.0),
-								alignment: Alignment.centerLeft,
-								child: new Row(
-									children: <Widget>[
-										new Expanded(
-											child: new Text('Header #$index',
-												style: const TextStyle(color: Colors.white),
-											),
-										),
-										new Offstage(
-											offstage: stuckAmount <= 0.0,
-											child: new Opacity(
-												opacity: stuckAmount,
-												child: new IconButton(
-													icon: new Icon(Icons.favorite, color: Colors.white),
-													onPressed: () =>
-														Scaffold.of(context).showSnackBar(
-															new SnackBar(content: new Text('Favorite #$index'))
-														),
-												),
-											),
-										),
-									],
-								),
-							);
-						},
-						content: new Container(
-							child: new Image.network(imageForIndex(index), fit: BoxFit.cover,
-								width: double.infinity, height: 200.0),
-						),
-					),
-				);
-			}),
-		);
-	}
+  const Example2({
+    Key? key,
+    this.controller,
+  }) : super(key: key);
 
-	String imageForIndex(int index) {
-		return Images.imageThumbUrls[index % Images.imageThumbUrls.length];
-	}
+  final ScrollController? controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return ScaffoldWrapper(
+      wrap: controller == null,
+      title: 'Example 2',
+      child: ListView.builder(
+        primary: controller == null,
+        controller: controller,
+        itemBuilder: (context, index) {
+          return Material(
+            color: Colors.grey[300],
+            child: StickyHeaderBuilder(
+              controller: controller, // Optional
+              builder: (BuildContext context, double stuckAmount) {
+                stuckAmount = 1.0 - stuckAmount.clamp(0.0, 1.0);
+                return Container(
+                  height: 50.0,
+                  color: Color.lerp(Colors.blue[700], Colors.red[700], stuckAmount),
+                  padding: EdgeInsets.symmetric(horizontal: 16.0),
+                  alignment: Alignment.centerLeft,
+                  child: Row(
+                    children: <Widget>[
+                      Expanded(
+                        child: Text(
+                          'Header #$index',
+                          style: const TextStyle(color: Colors.white),
+                        ),
+                      ),
+                      Offstage(
+                        offstage: stuckAmount <= 0.0,
+                        child: Opacity(
+                          opacity: stuckAmount,
+                          child: IconButton(
+                            icon: Icon(Icons.favorite, color: Colors.white),
+                            onPressed: () => ScaffoldMessenger.of(context)
+                                .showSnackBar(SnackBar(content: Text('Favorite #$index'))),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                );
+              },
+              content: Container(
+                child: Image.network(imageForIndex(index),
+                    fit: BoxFit.cover, width: double.infinity, height: 200.0),
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+
+  String imageForIndex(int index) {
+    return Images.imageThumbUrls[index % Images.imageThumbUrls.length];
+  }
 }
 
 class Example3 extends StatelessWidget {
-	@override
-	Widget build(BuildContext context) {
-		return new ScaffoldWrapper(
-			title: 'Example 3',
-			child: new ListView.builder(itemBuilder: (context, index) {
-				return new Material(
-					color: Colors.grey[300],
-					child: new StickyHeaderBuilder(
-						overlapHeaders: true,
-						builder: (BuildContext context, double stuckAmount) {
-							stuckAmount = 1.0 - stuckAmount.clamp(0.0, 1.0);
-							return new Container(
-								height: 50.0,
-								color: Colors.grey[900].withOpacity(0.6 + stuckAmount * 0.4),
-								padding: new EdgeInsets.symmetric(horizontal: 16.0),
-								alignment: Alignment.centerLeft,
-								child: new Text('Header #$index',
-									style: const TextStyle(color: Colors.white),
-								),
-							);
-						},
-						content: new Container(
-							child: new Image.network(imageForIndex(index), fit: BoxFit.cover,
-								width: double.infinity, height: 200.0),
-						),
-					),
-				);
-			}),
-		);
-	}
+  const Example3({
+    Key? key,
+    this.controller,
+  }) : super(key: key);
 
-	String imageForIndex(int index) {
-		return Images.imageThumbUrls[index % Images.imageThumbUrls.length];
-	}
+  final ScrollController? controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return ScaffoldWrapper(
+      wrap: controller == null,
+      title: 'Example 3',
+      child: ListView.builder(
+        primary: controller == null,
+        controller: controller,
+        itemBuilder: (context, index) {
+          return Material(
+            color: Colors.grey[300],
+            child: StickyHeaderBuilder(
+              overlapHeaders: true,
+              controller: controller, // Optional
+              builder: (BuildContext context, double stuckAmount) {
+                stuckAmount = 1.0 - stuckAmount.clamp(0.0, 1.0);
+                return Container(
+                  height: 50.0,
+                  color: Colors.grey.shade900.withOpacity(0.6 + stuckAmount * 0.4),
+                  padding: EdgeInsets.symmetric(horizontal: 16.0),
+                  alignment: Alignment.centerLeft,
+                  child: Text(
+                    'Header #$index',
+                    style: const TextStyle(color: Colors.white),
+                  ),
+                );
+              },
+              content: Container(
+                child: Image.network(
+                  imageForIndex(index),
+                  fit: BoxFit.cover,
+                  width: double.infinity,
+                  height: 200.0,
+                ),
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+
+  String imageForIndex(int index) {
+    return Images.imageThumbUrls[index % Images.imageThumbUrls.length];
+  }
 }
 
 class ScaffoldWrapper extends StatelessWidget {
-	final Widget child;
-	final String title;
+  const ScaffoldWrapper({
+    Key? key,
+    required this.title,
+    required this.child,
+    this.wrap = true,
+  }) : super(key: key);
 
-	const ScaffoldWrapper({
-		Key key,
-		@required this.title,
-		@required this.child,
-	}) : super(key: key);
+  final Widget child;
+  final String title;
+  final bool wrap;
 
-	@override
-	Widget build(BuildContext context) {
-		return new Scaffold(
-			appBar: new PreferredSize(
-				preferredSize: new Size.fromHeight(kToolbarHeight),
-				child: new Hero(
-					tag: 'app_bar',
-					child: new AppBar(
-						title: new Text(title),
-						elevation: 0.0,
-					),
-				),
-			),
-			body: child,
-		);
-	}
+  @override
+  Widget build(BuildContext context) {
+    if (wrap) {
+      return Scaffold(
+        appBar: PreferredSize(
+          preferredSize: Size.fromHeight(kToolbarHeight),
+          child: Hero(
+            tag: 'app_bar',
+            child: AppBar(
+              title: Text(title),
+              elevation: 0.0,
+            ),
+          ),
+        ),
+        body: child,
+      );
+    } else {
+      return Material(
+        child: child,
+      );
+    }
+  }
+}
+
+class Example4 extends StatefulWidget {
+  @override
+  _Example4State createState() => _Example4State();
+}
+
+class _Example4State extends State<Example4> {
+  late final _controller = List.generate(4, (_) => ScrollController());
+
+  @override
+  Widget build(BuildContext context) {
+    return Material(
+      child: DefaultTabController(
+        length: 3,
+        child: NestedScrollView(
+          controller: _controller[0],
+          headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
+            return <Widget>[
+              SliverAppBar(
+                title: const Text('Example 4'),
+                pinned: true,
+                forceElevated: innerBoxIsScrolled,
+                bottom: TabBar(
+                  tabs: <Tab>[
+                    Tab(text: 'Example 1'),
+                    Tab(text: 'Example 2'),
+                    Tab(text: 'Example 3'),
+                  ],
+                ),
+              ),
+            ];
+          },
+          body: MediaQuery.removePadding(
+            context: context,
+            removeTop: true,
+            child: TabBarView(
+              children: <Widget>[
+                Example1(controller: _controller[1]),
+                Example2(controller: _controller[2]),
+                Example3(controller: _controller[3]),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
 }
